alletrt

+++ src/app/admin/portals/content/content-list.component.ts
@@ -85,6 +85,8 @@ export class ContentComponent implements OnInit, OnDestroy {
 
   }
   contentDelete(id){
-    this.portalService.deleteContent(id)
+    if(confirm("Are you sure to delete "+id)) {
+      this.portalService.deleteContent(id)
+    }
   }
 }
